feat: Cleanup ios classes
