--- ios/Classes/FlutterUdidPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface FlutterUdidPlugin : NSObject<FlutterPlugin>
-@end
